shard aware: Move ShardInfo class into a Cython
since the computation was quite naive to begin with and in pure python
we could do much better with a calculation purely in C.
